python/examples/xbox_controller/requirements.txt
-f ../../../prebuilt
bosdyn-client >= 3.1

distro ; platform_system != "Windows"
XInput-Python ; platform_system == "Windows"
